fix: merge ExerciseTemplate into skill.py, remove file-level docstrings
