+++ backend/app/models/__init__.py
@@ -1,6 +1,5 @@
-from app.models.exercise_template import ExerciseTemplate
 from app.models.session import Attempt, Session
-from app.models.skill import Skill, SkillPrerequisite
+from app.models.skill import ExerciseTemplate, Skill, SkillPrerequisite
 from app.models.student import Student, StudentSkillState
 
 __all__ = [
