+++ backend/app/models/skill.py
@@ -1,15 +1,10 @@
-from __future__ import annotations
-
 import datetime
-from typing import TYPE_CHECKING
 
 from sqlalchemy import CheckConstraint, ForeignKey, String, func
+from sqlalchemy.dialects.postgresql import JSONB
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from app.database import Base
-
-if TYPE_CHECKING:
-    from app.models.exercise_template import ExerciseTemplate
 
 
 class SkillPrerequisite(Base):
@@ -18,6 +13,19 @@ class SkillPrerequisite(Base):
 
     skill_id: Mapped[str] = mapped_column(ForeignKey("skills.id"), primary_key=True)
     prerequisite_id: Mapped[str] = mapped_column(ForeignKey("skills.id"), primary_key=True)
+
+
+class ExerciseTemplate(Base):
+    __tablename__ = "exercise_templates"
+    __table_args__ = (CheckConstraint("difficulty BETWEEN 1 AND 3", name="valid_difficulty"),)
+
+    id: Mapped[str] = mapped_column(String(100), primary_key=True)
+    skill_id: Mapped[str] = mapped_column(ForeignKey("skills.id"))
+    difficulty: Mapped[int] = mapped_column(default=1)
+    template: Mapped[dict] = mapped_column(JSONB)
+    created_at: Mapped[datetime.datetime] = mapped_column(server_default=func.now())
+
+    skill: Mapped["Skill"] = relationship(back_populates="exercise_templates")
 
 
 class Skill(Base):
@@ -37,6 +45,6 @@ class Skill(Base):
         lazy="selectin",
     )
 
-    exercise_templates: Mapped[list["ExerciseTemplate"]] = relationship(
+    exercise_templates: Mapped[list[ExerciseTemplate]] = relationship(
         back_populates="skill", lazy="selectin"
     )
