+++ backend/app/skill_tree/tree.py
@@ -1,5 +1,3 @@
-"""Skill tree DAG: load, validate, and query the FWB skill tree."""
-
 from __future__ import annotations
 
 from pathlib import Path
